requirements.txt
Flask==2.2.5
matplotlib==3.8.0
numpy==2.2.1
Pillow==11.1.0
scikit_learn==1.2.2
seaborn==0.13.2
torch==2.3.0
torchvision==0.18.0
